move L8RequestObject and L8ResponseObject to separate crates
